# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API Terintegrasi menggunakan L5-Swagger",
        "contact": {
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server Environment"
        }
    ],
    "paths": {
...
        "/api/revenue": {
            "get": {
                "tags": [
                    "Revenue / Pendapatan"
                ],
                "summary": "Get Laporan Realisasi Dashboard",
                "description": "Mengambil rangkuman hitungan data realisasi pendapatan horizontal dan vertikal sesuai tabel format Excel berdasarkan tahun.",
                "operationId": "getRevenueDashboard",
                "parameters": [
                    {
                        "name": "tahun",
                        "in": "query",
                        "description": "Tahun laporan pendapatan yang ingin dicari",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1900
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get data berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get data berhasil"
                                        },
                                        "data": {
                                            "properties": {
                                                "tahun": {
                                                    "type": "integer",
                                                    "example": 1900
                                                },
                                                "target_tahunan": {
                                                    "type": "object",
                                                    "example": {
                                                        "UMUM": 2803200000,
                                                        "BPJS KESEHATAN": 284856000000
                                                    }
                                                },
                                                "total_target_tahunan": {
                                                    "type": "number",
                                                    "example": 330000000000
                                                },
                                                "target_bulanan": {
                                                    "type": "object",
                                                    "example": {
                                                        "UMUM": 233600000,
                                                        "BPJS KESEHATAN": 23738000000
                                                    }
                                                },
                                                "total_target_bulanan": {
                                                    "type": "number",
                                                    "example": 27500000000
                                                },
                                                "realisasi": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "bulan": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "bulan_name": {
                                                                "type": "string",
                                                                "example": "Januari"
                                                            },
                                                            "categories": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "category": {
                                                                            "type": "string",
                                                                            "example": "UMUM"
                                                                        },
                                                                        "amount": {
                                                                            "type": "number",
                                                                            "example": 2000000000
                                                                        },
                                                                        "percentage": {
                                                                            "type": "string",
                                                                            "example": "856.16"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            },
                                                            "total_bulan": {
                                                                "type": "number",
                                                                "example": 14800000000
                                                            },
                                                            "total_percentage": {
                                                                "type": "string",
                                                                "example": "57.09"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "summary": {
                                                    "properties": {
                                                        "total_per_kategori": {
                                                            "type": "object",
                                                            "example": {
                                                                "UMUM": 24000000000
                                                            }
                                                        },
                                                        "persentase_tahun": {
                                                            "type": "object",
                                                            "example": {
                                                                "UMUM": "856.16"
                                                            }
                                                        },
                                                        "grand_total_realisasi": {
                                                            "type": "number",
                                                            "example": 114738116031
                                                        },
                                                        "grand_persentase_tahun": {
                                                            "type": "string",
                                                            "example": "34.77"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request / Validasi Gagal"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Revenue / Pendapatan"
                ],
                "summary": "Store / Update Target & Realisasi Bulanan",
                "description": "Menyimpan atau memperbarui data master target tahunan, bulanan, beserta nominal cicilan realisasi pendapatan per kategori code (Khusus Superadmin).",
                "operationId": "storeOrUpdateRevenue",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tahun",
                                    "targets"
                                ],
                                "properties": {
                                    "tahun": {
                                        "type": "integer",
                                        "example": 1900
                                    },
                                    "targets": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "category_code",
                                                "target_tahunan",
                                                "target_bulanan"
                                            ],
                                            "properties": {
                                                "category_code": {
                                                    "type": "string",
                                                    "example": "umum"
                                                },
                                                "target_tahunan": {
                                                    "type": "number",
                                                    "example": 2803200000
                                                },
                                                "target_bulanan": {
                                                    "type": "number",
                                                    "example": 233600000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "realisasi": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "bulan",
                                                "categories"
                                            ],
                                            "properties": {
                                                "bulan": {
                                                    "type": "integer",
                                                    "maximum": 12,
                                                    "minimum": 1,
                                                    "example": 1
                                                },
                                                "categories": {
                                                    "type": "array",
                                                    "items": {
                                                        "required": [
                                                            "category_code",
                                                            "amount"
                                                        ],
                                                        "properties": {
                                                            "category_code": {
                                                                "type": "string",
                                                                "example": "umum"
                                                            },
                                                            "amount": {
                                                                "type": "number",
                                                                "example": 2000000000
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data target dan realisasi berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data target dan realisasi berhasil disimpan"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error (Unprocessable Entity)"
                    },
                    "401": {
                        "description": "Unauthenticated / Token missing"
                    },
                    "403": {
                        "description": "Forbidden (Bukan Superadmin)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/revenue/years": {
            "get": {
                "tags": [
                    "Revenue / Pendapatan"
                ],
                "summary": "Get List Tahun Terdaftar",
                "description": "Mengambil semua data tahun unik yang sudah terdaftar di database untuk kebutuhan dropdown opsi form filter (Khusus Superadmin).",
                "operationId": "getRevenueYears",
                "responses": {
                    "200": {
                        "description": "Get list tahun berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get list tahun berhasil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "example": 1900
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Anda tanpa kata 'Bearer '",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}